Migrate from old name to 'programatorus'
