+++ app/src/main/java/programatorus/client/WeakRefFactoryMixin.kt
@@ -1,0 +1,11 @@
+package programatorus.client
+
+import java.lang.ref.WeakReference
+
+interface WeakRefFactoryMixin<T> {
+
+    fun weakRefFromThis(): WeakReference<T> {
+        return WeakReference<T>(this as T)
+    }
+
+}
