+++ app/src/main/java/programatorus/client/transport/ConnectionState.kt
@@ -1,0 +1,13 @@
+package programatorus.client.transport
+
+enum class ConnectionState {
+    CONNECTING,
+
+    CONNECTED,
+
+    DISCONNECTING,
+
+    DISCONNECTED,
+
+    ERROR,
+}
