+++ app/src/main/java/programatorus/client/transport/IOutgoingMessage.kt
@@ -1,0 +1,12 @@
+package programatorus.client.transport
+
+import programus.proto.GenericMessage
+import java.util.concurrent.CompletableFuture
+
+interface IOutgoingMessage {
+
+    val message: GenericMessage
+
+    val response: CompletableFuture<GenericMessage>
+
+}
