+++ app/src/main/java/programatorus/client/transport/ITransportClient.kt
@@ -1,0 +1,14 @@
+package programatorus.client.transport
+
+import programus.proto.GenericMessage
+
+interface ITransportClient {
+
+    fun onStateChanged(state: ConnectionState) {}
+
+    // TODO(bgrzesik): Remove this in favor of onStateChanged(ConnectionState::ERROR)
+    fun onError() {}
+
+    fun onMessageReceived(message: GenericMessage) {}
+
+}
