--- app/src/main/java/programus/client/WeakRefFactoryMixin.kt
@@ -1,11 +1,0 @@
-package programus.client
-
-import java.lang.ref.WeakReference
-
-interface WeakRefFactoryMixin<T> {
-
-    fun weakRefFromThis(): WeakReference<T> {
-        return WeakReference<T>(this as T)
-    }
-
-}
