--- app/src/main/java/programus/client/transport/ConnectionState.kt
@@ -1,13 +1,0 @@
-package programus.client.transport
-
-enum class ConnectionState {
-    CONNECTING,
-
-    CONNECTED,
-
-    DISCONNECTING,
-
-    DISCONNECTED,
-
-    ERROR,
-}
