--- app/src/main/java/programus/client/transport/IOutgoingMessage.kt
@@ -1,12 +1,0 @@
-package programus.client.transport
-
-import programus.proto.GenericMessage
-import java.util.concurrent.CompletableFuture
-
-interface IOutgoingMessage {
-
-    val message: GenericMessage
-
-    val response: CompletableFuture<GenericMessage>
-
-}
