--- app/src/test/java/programus/client/transport/TransportTest.kt
@@ -1,101 +1,0 @@
-package programus.client.transport
-
-import android.util.Log
-import io.mockk.every
-import io.mockk.mockkStatic
-import org.junit.Assert
-import org.junit.Before
-import org.junit.Test
-import org.junit.jupiter.api.BeforeAll
-import programus.client.transport.io.LoopbackTransport
-import programus.client.transport.wrapper.Transport
-import programus.proto.GenericMessage
-import programus.proto.TestMessage
-import java.util.concurrent.ArrayBlockingQueue
-import java.util.concurrent.TimeUnit
-
-internal class TransportTest {
-
-    @Before
-    fun mockLog() = TestUtils.mockLog()
-
-    @Test
-    fun testLoopbackSendReceive1() {
-        testLoopbackSendReceive(1)
-    }
-
-    @Test
-    fun testLoopbackSendReceive10() {
-        testLoopbackSendReceive(10)
-    }
-
-    @Test
-    fun testLoopbackSendReceive100() {
-        testLoopbackSendReceive(100)
-    }
-
-    private fun testLoopbackSendReceive(num: Int) {
-        val queue = ArrayBlockingQueue<GenericMessage>(1)
-
-        val client = object : ITransportClient {
-            override fun onMessageReceived(message: GenericMessage) {
-                queue.add(message)
-            }
-
-            override fun onError() = Assert.fail()
-        }
-
-        val transport = Transport({ client, executor -> LoopbackTransport(client, executor) }, client)
-
-        val messages = mutableListOf<GenericMessage>()
-        for (i in 0..num) {
-            messages.add(TestUtils.newTestMessage())
-        }
-
-        for (message in messages) {
-            transport.send(message)
-
-            val received = queue.poll(100, TimeUnit.MILLISECONDS)
-
-            Assert.assertNotNull(received)
-            Assert.assertEquals(message, received)
-        }
-
-        transport.disconnect()
-    }
-
-    @Test
-    fun testStateMachine() {
-        val queue = ArrayBlockingQueue<ConnectionState>(1)
-
-        val client = object : ITransportClient {
-            override fun onStateChanged(state: ConnectionState) {
-                queue.add(state)
-            }
-
-            override fun onError() = Assert.fail()
-        }
-
-        val transport = Transport(::LoopbackTransport, client)
-
-        var state = queue.poll(100, TimeUnit.MILLISECONDS)
-        Assert.assertEquals(state, ConnectionState.DISCONNECTED)
-
-        transport.reconnect()
-
-        state = queue.poll(100, TimeUnit.MILLISECONDS)
-        Assert.assertEquals(state, ConnectionState.CONNECTING)
-
-        state = queue.poll(100, TimeUnit.MILLISECONDS)
-        Assert.assertEquals(state, ConnectionState.CONNECTED)
-
-        transport.disconnect()
-
-        state = queue.poll(100, TimeUnit.MILLISECONDS)
-        Assert.assertEquals(state, ConnectionState.DISCONNECTING)
-
-        state = queue.poll(100, TimeUnit.MILLISECONDS)
-        Assert.assertEquals(state, ConnectionState.DISCONNECTED)
-    }
-
-}
